toggle barra de accion, extracción panel de favoritos
